// packages/nest-api/src/scrapers/types/traderaListing.ts
export interface TraderaListing {
  currentBid: number;
  itemId: number;
  rootItemId: number;
  price: number;
  isAuction: boolean;
  isShopItem: boolean;
  shortDescription: string;
  imageUrl: string;
  imageHeight: number;
  imageWidth: number;
  itemUrl: string;
  itemType: "auction";
  totalBids: number;
  timeLeft: string;
  startDate: string;
  endDate: string;
  categoryId: number;
  buyNowPrice: number;
  sellerMemberId: number;
  showBids: boolean;
  showBuyNow: boolean;
  discountRate: number;
  priceBeforeDiscount: number;
  hasDiscount: boolean;
  canHaveDiscount: boolean;
  sellerAlias: string;
  showFreeShippingRibbon: boolean;
  showOnDisplayRibbon: boolean;
  sellerDsrAverage: number;
  prettifyRating: {
    rating: string;
    includeTotal: boolean;
    noRating: boolean;
    isGood: boolean;
    isBad: boolean;
  };
  isNewToday: boolean;
  isEndingSoon: boolean;
  isLessThanOneMinuteLeft: boolean;
  shippingPrice: string;
  isPickupOnly: boolean;
  hasNonSpecifiedShipping: boolean;
  showShipping: boolean;
  hasFreeShipping: boolean;
  existsInWishList: boolean;
  displayHighlighted: boolean;
  onDisplay: boolean;
  itemAttributesForFilter: [];
  sellerDsrAverageText: string;
  searchId: string;
  showShopItemSettings: boolean;
}